build our Helm chart
We will configure and deploy our service with Helm.

We've build a Helm chart for our service. We can install Helm chart
in our Kind (Kubernetes in Docker) cluster to run a cluster of
our service.

Helm charts include a template file for each resource type. Our
service require two resource types: a StatefulSet and a Service,
so we define resource we need in a `statefulset.yaml` file and
a `service.yaml` file.