fix container crashing
As we're still developing, we want to use alpine instead of the
scratch image for the final build to make it easier to
debug problems inside the running container.

We also created a `docker-entrypoint.sh` script so we can run
more commands to set up commitlog before we run the actual
binary. We need this to fix and workaround Kubernetes
container "CrashLoopBackOff" caused by Raft init problem/error:
"raft: Failed to commit logs: EOF ... log.*index.Write: EOF".

This problem happened in situation where we set Kubernetes
readinessProbe initialDelaySeconds to a value lower than the
time taken to init Raft and start commitlog. If this condition is
true, then Kubernetes will restart the unhealthy pod/container.
The next time the container start, our data directory is
not empty and because of this, commitlog's Raft doesn't
deal with this case well and hence we're seeing the Raft's "EOF"
error.